feat(data-quality): add file type filter to archives hook and intake files table
- Add optional `fileType` parameter to `useArchivesClientData` hook to filter archives by file type.
- Create `IntakeFilesTable` component with file type selection dropdown and upload modal trigger.
- Integrate `useArchivesClientData` and `useFileTypes` hooks to display filtered archives.
